requirements.txt
aiogram==3.13.1
openai==1.54.3
pydub==0.25.1
python-dotenv==1.0.0
aiohttp==3.10.11

